+++ src/main/java/roomescape/admin/dto/AdminReservationRequest.java
@@ -1,0 +1,10 @@
+package roomescape.admin.dto;
+
+import java.time.LocalDate;
+import roomescape.reservation.dto.ReservationRequest;
+
+public record AdminReservationRequest(LocalDate date, Long timeId, Long themeId, Long memberId) {
+    public ReservationRequest getReservationRequest(){
+        return new ReservationRequest(date,timeId,themeId);
+    }
+}
