+++ src/main/java/roomescape/admin/presentation/AdminController.java
@@ -1,30 +1,36 @@
 package roomescape.admin.presentation;
 
-import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
+import static roomescape.admin.presentation.AdminController.ADMIN_BASE_URL;
+import static roomescape.member.presentation.MemberController.RESERVATION_BASE_URL;
 
-@Controller
-@RequestMapping("/admin")
+import java.net.URI;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+import roomescape.admin.dto.AdminReservationRequest;
+import roomescape.reservation.dto.ReservationResponse;
+import roomescape.reservation.service.ReservationService;
+
+@RestController
+@RequestMapping(ADMIN_BASE_URL)
 public class AdminController {
 
-    @GetMapping
-    public String admin() {
-        return "admin/index";
+    private final ReservationService reservationService;
+
+    public static final String ADMIN_BASE_URL = "/admin";
+    private static final String SLASH = "/";
+
+    public AdminController(ReservationService reservationService) {
+        this.reservationService = reservationService;
     }
 
-    @GetMapping("/reservation")
-    public String reservation() {
-        return "admin/reservation-new";
-    }
-
-    @GetMapping("/time")
-    public String time() {
-        return "admin/time";
-    }
-
-    @GetMapping("/theme")
-    public String theme() {
-        return "admin/theme";
+    @PostMapping("/reservations")
+    public ResponseEntity<ReservationResponse> createReservation(@RequestBody final AdminReservationRequest request) {
+        ReservationResponse response = reservationService.createReservation(request.getReservationRequest(),
+                request.memberId());
+        URI locationUri = URI.create(RESERVATION_BASE_URL + SLASH + response.id());
+        return ResponseEntity.created(locationUri).body(response);
     }
 }
